requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
